feat: replace admin pages with link to Supabase
