+++ public/static/locales/th/admin.json
@@ -1,13 +1,5 @@
 {
   "title": "แผงควบคุม",
-  "suggestions": {
-    "title": "ข้อแนะนำ",
-    "newYearNewData": {
-      "title": "ปีใหม่ ข้อมูลใหม่",
-      "desc": "เราสังเกตว่าบางตารางในปี {{year}} ว่างเปล่า ทำไมไม่เริ่มต้นด้วยการเพิ่มข้อมูลสำหรับปีนี้",
-      "contactSupport": "ติดต่อสำหรับรายละเอียดเพิ่มเติม"
-    }
-  },
   "statistics": {
     "student": {
       "all": "นักเรียน {{count, number}} คน",
@@ -27,113 +19,10 @@
     },
     "lastUpdated": "ข้อมูล ณ {{date}}"
   },
-  "data": {
-    "title": "ระบบข้อมูล",
-    "manage": {
-      "title": "ดูและแก้ข้อมูลของโรงเรียน",
-      "desc": "เข้าถึงข้อมูลทั้งหมดของโรงเรียน สามารถเพิ่ม แก้ไข ดู และลบรายการได้",
-      "action": {
-        "student": "นักเรียน",
-        "teacher": "อาจารย์",
-        "subject": "วิชา",
-        "class": "ห้องเรียน"
-      },
-      "student": {
-        "title": "จัดการข้อมูลนักเรียน",
-        "thead": {
-          "studentID": "รหัสนักเรียน",
-          "nameTH": "ชื่อ-นามสกุล",
-          "nameEN": "ชื่อ-นามสกุลภาษาอังกฤษ",
-          "class": "ห้องเรียน",
-          "classNo": "เลขที่ในห้อง"
-        }
-      },
-      "subject": {
-        "title": "จัดการข้อมูลวิชา",
-        "thead": {
-          "codeTH": "รหัสวิชา",
-          "codeEN": "รหัสวิชาภาษาอังกฤษ",
-          "nameTH": "วิชา",
-          "nameEN": "วิชาภาษาอังกฤษ",
-          "year": "ปีการศึกษา",
-          "semester": "ภาคเรียน"
-        }
-      },
-      "class": {
-        "title": "จัดการข้อมูลห้องเรียน",
-        "thead": {
-          "number": "ห้อง",
-          "classAdvisorsTH": "ครูที่ปรึกษา",
-          "classAdvisorsEN": "ครูที่ปรึกษา (ภาษาอังกฤษ)",
-          "studentCount": "จำนวนนักเรียน",
-          "year": "ปีการศึกษา"
-        },
-        "tbody": {
-          "class": "ม.{{number}}",
-          "studentCount_zero": "ไม่มีนักเรียน",
-          "studentCount_other": "{{count, number}} คน"
-        }
-      }
-    },
-    "import": {
-      "title": "นำเข้าข้อมูลด้วย CSV",
-      "desc": "สร้างไฟล์ CSV ในรูปแบบที่เหมาะสมด้วยซอฟต์แวร์สเปรดชีต เช่น Google Sheet หรือ Microsoft Excel แล้วนำเข้าที่นี่",
-      "action": {
-        "student": "นำเข้านักเรียน",
-        "teacher": "นำเข้าอาจารย์",
-        "subject": "นำเข้าวิชา",
-        "class": "สร้างห้องเรียน"
-      },
-      "dialog": {
-        "import": {
-          "title": {
-            "student": "นำเข้านักเรียน",
-            "teacher": "นำเข้าอาจารย์",
-            "subject": "นำเข้าวิชา"
-          },
-          "desc": "ผู้ดูแลระบบสามารนำเข้าไฟล์ CSV สู่ฐานข้อมูล MySK กรุณาตรวจสอบให้แน่ใจว่าไฟล์ CSV ที่กำลังจะนำเข้ามีหัวตารางดังต่อไปนี้",
-          "table": {
-            "header": "หัวตาราง",
-            "contentType": "ประเภทข้อมูล",
-            "required": "ต้องมีข้อมูลทุกแถว"
-          },
-          "form": {
-            "csvFile": "ไฟล์ CSV ที่จะนำเข้า",
-            "hasHeader": "ไฟล์นี้มีหัวตาราง"
-          },
-          "action": {
-            "import": "นำเข้า",
-            "downloadStarter": "ดาวน์โหลดไฟล์เริ่มต้น"
-          }
-        },
-        "generateClasses": {
-          "title": "สร้างห้องเรียนอัตโนมัติ",
-          "desc": "สร้างทุกห้องเรียนสำหรับปีการศึกษานี้ด้วยเครื่องมือนี้",
-          "numGrades": "จำนวนระดับ",
-          "numClasses": {
-            "title": "จำนวนห้องต่อระดับ",
-            "instructions": "ใส่จำนวนห้องเรียนที่ต้องการในแต่ละระดับ ตัวอย่างเช่น หากต้องการห้องเรียน <0>ม.101, ม.102, ม.103</0> ให้ป้อน “3” ในช่องที่ระบุว่า “จำนวนห้องในม.1”",
-            "input": "จำนวนห้องในม.{{grade}}"
-          },
-          "preview": {
-            "title": "ผลลัพธ์",
-            "class": "ม.{{number}}"
-          },
-          "action": {
-            "cancel": "ยกเลิก",
-            "generate": "สร้างอัตโนมัติ"
-          }
-        }
-      }
-    }
-  },
-  "news": {
-    "title": "โต้ตอบกับนักเรียนด้วยบทความข่าวสารและแบบฟอร์ม",
-    "desc": "คุณสามารถสร้างบทความด้วยข้อความ รูปภาพ และตาราง ที่ปรากฏภายใน MySK เพื่อนักเรียนและครู สำหรับการรวบรวมข้อมูล คุณสามารถสร้างแบบฟอร์มให้นักเรียนกรอก และอาจารย์ที่ปรึกษาประจำชั้นเรียนสามารถดูผลลัพธ์ได้",
+  "supabase": {
+    "title": "จัดการข้อมูลด้วย Supabase Table Editor",
     "action": {
-      "createArticle": "สร้างบทความข่าวสาร",
-      "createForm": "สร้างแบบฟอร์ม",
-      "manage": "จัดการที่มีอยู่แล้ว"
+      "open": "เปิด Supabase"
     }
   }
 }
